services/memory-mcp/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
jsonschema==4.23.0
